requirements.txt
asgiref==3.7.2
Django==4.2.4
django-htmx==1.17.0
django-redis==5.4.0
plum-dispatch==2.2.2
waitress==2.1.2
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
python-interface==1.6.1
